requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
pdf2image==1.17.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
python-dotenv==1.1.1
requests==2.32.4
sniffio==1.3.1
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
python-multipart==0.0.20
